+++ androidshared/src/main/java/org/odk/collect/androidshared/ui/MultiSelectViewModel.kt
@@ -1,6 +1,8 @@
 package org.odk.collect.androidshared.ui
 
+import android.widget.Button
 import androidx.lifecycle.ViewModel
+import org.odk.collect.androidshared.R
 import org.odk.collect.androidshared.livedata.MutableNonNullLiveData
 import org.odk.collect.androidshared.livedata.NonNullLiveData
 
@@ -31,4 +33,16 @@ class MultiSelectViewModel : ViewModel() {
             select(item)
         }
     }
+}
+
+fun updateSelectAll(button: Button, itemCount: Int, selectedCount: Int): Boolean {
+    val allSelected = itemCount > 0 && selectedCount == itemCount
+
+    if (allSelected) {
+        button.setText(R.string.clear_all)
+    } else {
+        button.setText(R.string.select_all)
+    }
+
+    return allSelected
 }
